wasm confirmed working
