+++ src/Blazor.FileReader.Wasm.Demo/App.cshtml
