+++ src/Blazor.FileReader.Wasm.Demo/Blazor.FileReader.Wasm.Demo.csproj
