+++ src/Blazor.FileReader.Wasm.Demo/Pages/Assembly.cshtml
