+++ src/Blazor.FileReader.Wasm.Demo/Pages/SHA256.cshtml
