+++ src/Blazor.FileReader.Wasm.Demo/Pages/_ViewImports.cshtml
