+++ src/Blazor.FileReader.Wasm.Demo/Shared/MainLayout.cshtml
