+++ src/Blazor.FileReader.Wasm.Demo/Shared/NavMenu.cshtml
