+++ src/Blazor.FileReader.Wasm.Demo/_ViewImports.cshtml
@@ -3,7 +3,7 @@
 @using Microsoft.AspNetCore.Blazor.Components
 @using Microsoft.AspNetCore.Blazor.Layouts
 @using Microsoft.AspNetCore.Blazor.Routing
-@using Blazor.FileReader.Demo
-@using Blazor.FileReader.Demo.Shared
+@using Blazor.FileReader.Wasm.Demo
+@using Blazor.FileReader.Wasm.Demo.Shared
 @using Blazor.FileReader
 @addTagHelper *, Blazor.FileReader
