+++ src/Blazor.FileReader.Wasm.Demo/global.json
