+++ src/Blazor.FileReader.Wasm.Demo/wwwroot/css/open-iconic/FONT-LICENSE
