+++ src/Blazor.FileReader.Wasm.Demo/wwwroot/css/open-iconic/README.md
