+++ src/Blazor.FileReader.Wasm.Demo/wwwroot/css/open-iconic/font/fonts/open-iconic.woff
