+++ src/Blazor.FileReader.Wasm.Demo/wwwroot/index.html
