+++ src/Blazor.FileReader.sln
@@ -3,13 +3,13 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.27705.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Blazor.FileReader.Wasm.Demo", "Blazor.FileReader.Demo\Blazor.FileReader.Wasm.Demo.csproj", "{0781BC55-3054-47ED-A6B2-37448CA58038}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Blazor.FileReader", "Blazor.FileReader\Blazor.FileReader.csproj", "{5F8084DD-9D88-4407-BCFC-15C71C5AF11E}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Blazor.FileReader.AspNetCore.Demo.Server", "Blazor.FileReader.AspNetCore.Demo.Server\Blazor.FileReader.AspNetCore.Demo.Server.csproj", "{0C7A0ECA-735A-4C82-BECD-4730C2244B65}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Blazor.FileReader.AspNetCore.Demo.Server", "Blazor.FileReader.AspNetCore.Demo.Server\Blazor.FileReader.AspNetCore.Demo.Server.csproj", "{0C7A0ECA-735A-4C82-BECD-4730C2244B65}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Blazor.FileReader.AspNetCore.Demo.App", "Blazor.FileReader.AspNetCore.Demo.App\Blazor.FileReader.AspNetCore.Demo.App.csproj", "{EB8D9B98-1596-4504-B19A-DA090CC8AC76}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Blazor.FileReader.AspNetCore.Demo.App", "Blazor.FileReader.AspNetCore.Demo.App\Blazor.FileReader.AspNetCore.Demo.App.csproj", "{EB8D9B98-1596-4504-B19A-DA090CC8AC76}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Blazor.FileReader.Wasm.Demo", "Blazor.FileReader.Wasm.Demo\Blazor.FileReader.Wasm.Demo.csproj", "{9FA01C57-BBB5-420E-AA9B-AD22F39B77ED}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -18,12 +18,6 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{0781BC55-3054-47ED-A6B2-37448CA58038}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{0781BC55-3054-47ED-A6B2-37448CA58038}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{0781BC55-3054-47ED-A6B2-37448CA58038}.Ghpages|Any CPU.ActiveCfg = Ghpages|Any CPU
-		{0781BC55-3054-47ED-A6B2-37448CA58038}.Ghpages|Any CPU.Build.0 = Ghpages|Any CPU
-		{0781BC55-3054-47ED-A6B2-37448CA58038}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{0781BC55-3054-47ED-A6B2-37448CA58038}.Release|Any CPU.Build.0 = Release|Any CPU
 		{5F8084DD-9D88-4407-BCFC-15C71C5AF11E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{5F8084DD-9D88-4407-BCFC-15C71C5AF11E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5F8084DD-9D88-4407-BCFC-15C71C5AF11E}.Ghpages|Any CPU.ActiveCfg = Ghpages|Any CPU
@@ -41,6 +35,12 @@ Global
 		{EB8D9B98-1596-4504-B19A-DA090CC8AC76}.Ghpages|Any CPU.Build.0 = Debug|Any CPU
 		{EB8D9B98-1596-4504-B19A-DA090CC8AC76}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{EB8D9B98-1596-4504-B19A-DA090CC8AC76}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9FA01C57-BBB5-420E-AA9B-AD22F39B77ED}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9FA01C57-BBB5-420E-AA9B-AD22F39B77ED}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9FA01C57-BBB5-420E-AA9B-AD22F39B77ED}.Ghpages|Any CPU.ActiveCfg = Ghpages|Any CPU
+		{9FA01C57-BBB5-420E-AA9B-AD22F39B77ED}.Ghpages|Any CPU.Build.0 = Ghpages|Any CPU
+		{9FA01C57-BBB5-420E-AA9B-AD22F39B77ED}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9FA01C57-BBB5-420E-AA9B-AD22F39B77ED}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
