+++ src/Blazor.FileReader/BrowserFileReaderException.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace Blazor.FileReader
+{
+    public class BrowserFileReaderException : Exception
+    {
+        public BrowserFileReaderException(string message):base(message)
+        {
+        }
+    }
+}
