+++ src/Blazor.FileReader/ReadFileAsyncCallbackParams.cs
@@ -1,0 +1,11 @@
+﻿namespace Blazor.FileReader
+{
+    public partial class FileReaderJsInterop
+    {
+        public class ReadFileAsyncCallbackParams
+        {
+            public long CallBackId { get; set; }
+            public long BytesRead { get; set; }
+        }
+    }
+}
